Update mutation table and create draft of JSON file

+++ flask-server/server.py
@@ -4,6 +4,7 @@ import openai
 import config
 import enzy_htp.structure
 import enzy_htp.mutation.api as mapi
+import enzy_htp.mutation.mutation as mt
 import enzy_htp.mutation.mutation_pattern.api as pattern_api
 from enzy_htp.preparation import validity as vd
 from flask import Flask
@@ -13,7 +14,10 @@ from enzy_htp.core import _LOGGER
 
 app = Flask(__name__)
 CORS(app)
+
+# TODO: Update method for getting files - global variable for now, integrate into database
 app.config['UPLOAD_FOLDER'] = 'uploads'
+file_path = ""
 
 # import settings
 # app = Flask(__name__)
@@ -23,12 +27,6 @@ app.config['UPLOAD_FOLDER'] = 'uploads'
 # login_manager.login_message_category = "info"
 
 
-
-# # Example API route - to start server, run "python server.py"
-# # @app.route("/members")
-# # def members():
-# #     return {"members": ["Member1", "Member2"]}
-
 # Import and define your routes and views
 from auth import auth as auth_blueprint
 app.register_blueprint(auth_blueprint, url_prefix='/api/auth')
@@ -36,6 +34,7 @@ app.register_blueprint(auth_blueprint, url_prefix='/api/auth')
 # Validate File
 @app.route("/api/validate_file", methods=["POST"])
 def validate_file():
+    global file_path
     result = None
     message = """"""
     is_valid = None
@@ -59,19 +58,22 @@ def validate_file():
             message += "The following errors were found in the PDB file: \n"
             for reason, source, suggestion in intermediate_message:
                 message += "Reason: " + str(reason) + " Source: " + str(source) + " Suggestion: " + str(suggestion) + "\n"
-            os.remove(file_path)
+            if not is_valid:
+                os.remove(file_path)
         except:
             is_valid = False
             message = log_str.getvalue()
             os.remove(file_path)
+
     return jsonify({"validity": is_valid, "message": message})
 
 # Generate Patterns
 @app.route("/api/generate_pattern", methods=["POST"])
 def generate_pattern():
+    global file_path
     data = request.json
     mutation_request = data.get('mut_request')
-    api_key = request.json['api_key']
+    api_key = data.get('api_key')
 
     prompt = ""
     prompt += config.prompt_skeleton
@@ -80,19 +82,37 @@ def generate_pattern():
     openai.api_key = api_key
 
     # TODO: how to improve prompt in config.py?
+    # try:
+    #     completions = openai.Completion.create(
+    #         engine="gpt-3.5-turbo-instruct",
+    #         prompt=prompt,
+    #         max_tokens=70,
+    #         frequency_penalty=-0.5,
+    #         temperature=0.01,
+    #     )
+    #     pattern = completions.choices[0].text
+    # except Exception as e:
+    #     raise Exception(f'API Error: {str(e)}')
+    
+    pattern = "r:3[resi 1 around 4:all not self]*10"
+    sp = enzy_htp.structure.PDBParser()
+    stru = sp.get_structure(file_path)
     try:
-        completions = openai.Completion.create(
-            engine="gpt-3.5-turbo-instruct",
-            prompt=prompt,
-            max_tokens=70,
-            frequency_penalty=-0.5,
-            temperature=0.01,
-        )
-        message = completions.choices[0].text
-    except Exception as e:
-        raise Exception(f'API Error: {str(e)}')
+        mutations = pattern_api.decode_mutation_pattern(stru, pattern)
+    except pattern_api.InvalidMutationPatternSyntax as e:
+        raise Exception(f'Invalid mutation: {str(e)}')
+    
+    mut_string = ""
+    for mut in mutations:
+        mut_string += mt.get_mutant_name_str(mut) + ";"
+    
+    mut_string = mut_string[:-1]
 
-    return jsonify({"mutations": message})
+    os.remove(file_path)
+
+    # TODO: generate mutants with "generate_mut" function and save it
+
+    return jsonify({"pattern": pattern, "mut_string": mut_string})
 
 def generate_muts(file, pattern):
     sp = enzy_htp.structure.PDBParser()
@@ -112,11 +132,8 @@ def generate_muts(file, pattern):
             res_file = sp.get_file_str(mutant_stru)
         except Exception as e:
             raise Exception(f'API Error: {str(e)}')
-        name_tag = ""
-        for single_mut in mut:
-            name_tag += str(single_mut) + "_"
-        name_tag = name_tag[:-1]
-        res.append((res_file, name_tag))
+    name_tag = mt.get_mutant_name_tag(mutations)
+    res.append((res_file, name_tag))
 
     return res
